+++ adapters/AppLovin/AppLovinAdapter/GADMAdapterAppLovinConstant.m
@@ -14,8 +14,12 @@
   return @"com.applovin.sdk.mediation.admob.errorDomain";
 }
 
++ (NSString *)rtbErrorDomain {
+  return @"com.applovin.sdk.mediation.admob.rtb.errorDomain";
+}
+
 + (NSString *)adapterVersion {
-  return @"5.1.1.0";
+  return @"6.6.1.0";
 }
 
 + (NSString *)sdkKey {