+++ adapters/MoPub/MoPubAdapter/MoPubAdapterConstants.h
@@ -1,3 +1,15 @@
 #define MINIMUM_MOPUB_PRIVACY_ICON_SIZE 10
 #define DEFAULT_MOPUB_PRIVACY_ICON_SIZE 20
 #define MAXIMUM_MOPUB_PRIVACY_ICON_SIZE 30
+
+/// MoPub mediation network adapter version.
+static NSString *const kGADMAdapterMoPubVersion = @"5.6.0.1";
+
+/// Constant for adapter error domain.
+static NSString *const kGADMAdapterMoPubErrorDomain = @"com.mopub.mobileads.mopubadapter";
+
+/// Internal to MoPub
+static NSString *const kGADMAdapterMoPubTpValue = @"gmext";
+
+/// MoPub Pub id Key
+static NSString *const kGADMAdapterMoPubPubIdKey = @"pubid";