+++ adapters/Unity/UnityAdapter/GADMAdapterUnity.m
@@ -16,88 +16,49 @@
 
 #import "GADMAdapterUnityConstants.h"
 #import "GADMAdapterUnitySingleton.h"
+#import "GADMediationAdapterUnity.h"
 #import "GADUnityError.h"
 
 @interface GADMAdapterUnity () {
   /// Connector from Google Mobile Ads SDK to receive ad configurations.
-  __weak id<GADMRewardBasedVideoAdNetworkConnector> _rewardBasedVideoAdConnector;
+  __weak id<GADMAdNetworkConnector> _networkConnector;
 
-  /// Connector from Google Mobile Ads SDK to receive ad configurations.
-  __weak id<GADMAdNetworkConnector> _interstitialConnector;
+  /// The Requested Banner Ad size.
+  GADAdSize _requestedAdSize;
+
+  /// Game ID of Unity Ads network.
+  NSString *_gameID;
 
   /// Placement ID of Unity Ads network.
   NSString *_placementID;
 
   /// YES if the adapter is loading.
   BOOL _isLoading;
+
+  /// YES if a UnityAds Banner has loaded.
+  BOOL _bannerDidLoad;
 }
 
 @end
 
 @implementation GADMAdapterUnity
 
++ (nonnull Class<GADMediationAdapter>)mainAdapterClass {
+  return [GADMediationAdapterUnity class];
+}
+
 + (NSString *)adapterVersion {
-  return GADMAdapterUnityVersion;
+  return kGADMAdapterUnityVersion;
 }
 
 + (Class<GADAdNetworkExtras>)networkExtrasClass {
   return Nil;
 }
 
-#pragma mark Reward-based Video Ad Methods
-
-- (instancetype)initWithRewardBasedVideoAdNetworkConnector:
-        (id<GADMRewardBasedVideoAdNetworkConnector>)connector {
-  if (!connector) {
-    return nil;
-  }
-
-  self = [super init];
-  if (self) {
-    _rewardBasedVideoAdConnector = connector;
-  }
-  return self;
-}
-
-- (void)setUp {
-  NSString *gameID =
-      [[[_rewardBasedVideoAdConnector credentials] objectForKey:GADMAdapterUnityGameID] copy];
-  _placementID =
-      [[[_rewardBasedVideoAdConnector credentials] objectForKey:GADMAdapterUnityPlacementID] copy];
-  if (!gameID || !_placementID) {
-    NSError *error = GADUnityErrorWithDescription(@"Game ID and Placement ID cannot be nil.");
-    [_rewardBasedVideoAdConnector adapter:self didFailToSetUpRewardBasedVideoAdWithError:error];
-    return;
-  }
-  BOOL isConfigured =
-      [[GADMAdapterUnitySingleton sharedInstance] configureRewardBasedVideoAdWithGameID:gameID
-                                                                               delegate:self];
-  if (isConfigured) {
-    [_rewardBasedVideoAdConnector adapterDidSetUpRewardBasedVideoAd:self];
-  } else {
-    NSString *description =
-        [[NSString alloc] initWithFormat:@"%@ is not supported for this device.",
-                                         NSStringFromClass([UnityAds class])];
-    NSError *error = GADUnityErrorWithDescription(description);
-    [_rewardBasedVideoAdConnector adapter:self didFailToSetUpRewardBasedVideoAdWithError:error];
-  }
-}
-
-- (void)requestRewardBasedVideoAd {
-  _isLoading = YES;
-  [[GADMAdapterUnitySingleton sharedInstance] requestRewardBasedVideoAdWithDelegate:self];
-}
-
-- (void)presentRewardBasedVideoAdWithRootViewController:(UIViewController *)viewController {
-  // We will send adapterDidOpenRewardBasedVideoAd callback before presenting the unity ad because
-  // the ad has already loaded.
-  [_rewardBasedVideoAdConnector adapterDidOpenRewardBasedVideoAd:self];
-  [[GADMAdapterUnitySingleton sharedInstance]
-      presentRewardBasedVideoAdForViewController:viewController
-                                        delegate:self];
-}
-
 - (void)stopBeingDelegate {
+  if (_bannerDidLoad) {
+    [UnityAdsBanner destroy];
+  }
   [[GADMAdapterUnitySingleton sharedInstance] stopTrackingDelegate:self];
 }
 
@@ -110,30 +71,28 @@
 
   self = [super init];
   if (self) {
-    _interstitialConnector = connector;
+    _networkConnector = connector;
   }
   return self;
 }
 
 - (void)getInterstitial {
-  NSString *gameID =
-      [[[_interstitialConnector credentials] objectForKey:GADMAdapterUnityGameID] copy];
-  _placementID =
-      [[[_interstitialConnector credentials] objectForKey:GADMAdapterUnityPlacementID] copy];
-  if (!gameID || !_placementID) {
+  id<GADMAdNetworkConnector> strongConnector = _networkConnector;
+  _gameID = [[[strongConnector credentials] objectForKey:kGADMAdapterUnityGameID] copy];
+  _placementID = [[[strongConnector credentials] objectForKey:kGADMAdapterUnityPlacementID] copy];
+  if (!_gameID || !_placementID) {
     NSError *error = GADUnityErrorWithDescription(@"Game ID and Placement ID cannot be nil.");
-    [_interstitialConnector adapter:self didFailAd:error];
+    [strongConnector adapter:self didFailAd:error];
     return;
   }
   _isLoading = YES;
-  [[GADMAdapterUnitySingleton sharedInstance] configureInterstitialAdWithGameID:gameID
-                                                                       delegate:self];
+  [[GADMAdapterUnitySingleton sharedInstance] requestInterstitialAdWithDelegate:self];
 }
 
 - (void)presentInterstitialFromRootViewController:(UIViewController *)rootViewController {
   // We will send adapterWillPresentInterstitial callback before presenting unity ad because the ad
   // has already loaded.
-  [_interstitialConnector adapterWillPresentInterstitial:self];
+  [_networkConnector adapterWillPresentInterstitial:self];
   [[GADMAdapterUnitySingleton sharedInstance]
       presentInterstitialAdForViewController:rootViewController
                                     delegate:self];
@@ -142,10 +101,27 @@
 #pragma mark Banner Methods
 
 - (void)getBannerWithSize:(GADAdSize)adSize {
-  // Unity Ads doesn't support banner ads.
-  id<GADMAdNetworkConnector> strongConnector = _interstitialConnector;
-  NSError *error = GADUnityErrorWithDescription(@"Unity Ads doesn't support banner ads.");
-  [strongConnector adapter:self didFailAd:error];
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  _requestedAdSize = [self supportedAdSizeFromRequestedSize:adSize];
+
+  if (!IsGADAdSizeValid(_requestedAdSize)) {
+    NSLog(@"Requested unsupported banner size: %@", NSStringFromGADAdSize(adSize));
+    NSError *error = GADUnityErrorWithDescription(@"Requested unsupported banner size.");
+    [strongNetworkConnector adapter:self didFailAd:error];
+    return;
+  }
+
+  _gameID = [[[strongNetworkConnector credentials] objectForKey:kGADMAdapterUnityGameID] copy];
+  _placementID =
+      [[[strongNetworkConnector credentials] objectForKey:kGADMAdapterUnityPlacementID] copy];
+  if (!_gameID || !_placementID) {
+    NSError *error = GADUnityErrorWithDescription(@"Game ID and Placement ID cannot be nil.");
+    [strongNetworkConnector adapter:self didFailAd:error];
+    return;
+  }
+
+  _bannerDidLoad = NO;
+  [[GADMAdapterUnitySingleton sharedInstance] presentBannerAd:_gameID delegate:self];
 }
 
 - (BOOL)isBannerAnimationOK:(GADMBannerAnimationType)animType {
@@ -154,8 +130,20 @@
 
 #pragma mark GADMAdapterUnityDataProvider Methods
 
+- (NSString *)getGameID {
+  return _gameID;
+}
+
 - (NSString *)getPlacementID {
   return _placementID;
+}
+
+/// Find closest supported ad size from a given ad size.
+/// Returns nil if no supported size matches.
+- (GADAdSize)supportedAdSizeFromRequestedSize:(GADAdSize)gadAdSize {
+  NSArray *potentials =
+      @[ NSValueFromGADAdSize(kGADAdSizeBanner), NSValueFromGADAdSize(kGADAdSizeLeaderboard) ];
+  return GADClosestValidSizeForAdSizes(gadAdSize, potentials);
 }
 
 #pragma mark - Unity Delegate Methods
@@ -169,78 +157,117 @@
 }
 
 - (void)unityAdsDidFinish:(NSString *)placementID withFinishState:(UnityAdsFinishState)state {
-  if (_interstitialConnector) {
-    [_interstitialConnector adapterWillDismissInterstitial:self];
-    [_interstitialConnector adapterDidDismissInterstitial:self];
-  } else if (_rewardBasedVideoAdConnector) {
-    if (state == kUnityAdsFinishStateCompleted) {
-      [_rewardBasedVideoAdConnector adapterDidCompletePlayingRewardBasedVideoAd:self];
-      // Unity Ads doesn't provide a way to set the reward on their front-end. Default to a reward
-      // amount of 1. Publishers using this adapter should override the reward on the AdMob
-      // front-end.
-      GADAdReward *reward =
-          [[GADAdReward alloc] initWithRewardType:@"" rewardAmount:[NSDecimalNumber one]];
-      [_rewardBasedVideoAdConnector adapter:self didRewardUserWithReward:reward];
-    }
-    [_rewardBasedVideoAdConnector adapterDidCloseRewardBasedVideoAd:self];
-  }
-}
-
-- (void)unityAdsDidStart:(NSString *)placementID {
-  if (_rewardBasedVideoAdConnector) {
-    [_rewardBasedVideoAdConnector adapterDidStartPlayingRewardBasedVideoAd:self];
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterWillDismissInterstitial:self];
+    [strongNetworkConnector adapterDidDismissInterstitial:self];
   }
 }
 
 - (void)unityAdsReady:(NSString *)placementID {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
   if (!_isLoading) {
     return;
   }
 
-  if (_interstitialConnector) {
-    [_interstitialConnector adapterDidReceiveInterstitial:self];
-  } else {
-    [_rewardBasedVideoAdConnector adapterDidReceiveRewardBasedVideoAd:self];
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterDidReceiveInterstitial:self];
   }
   _isLoading = NO;
 }
 
 - (void)unityAdsDidClick:(NSString *)placementID {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
   // The Unity Ads SDK doesn't provide an event for leaving the application, so the adapter assumes
   // that a click event indicates the user is leaving the application for a browser or deeplink, and
   // notifies the Google Mobile Ads SDK accordingly.
-  if (_interstitialConnector) {
-    [_interstitialConnector adapterDidGetAdClick:self];
-    [_interstitialConnector adapterWillLeaveApplication:self];
-  } else {
-    [_rewardBasedVideoAdConnector adapterDidGetAdClick:self];
-    [_rewardBasedVideoAdConnector adapterWillLeaveApplication:self];
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterDidGetAdClick:self];
+    [strongNetworkConnector adapterWillLeaveApplication:self];
   }
 }
 
 - (void)unityAdsDidError:(UnityAdsError)error withMessage:(NSString *)message {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
   if (!_isLoading) {
     // Unity Ads show error will only happen after the ad has been loaded. So, we will send
     // dismiss/close callbacks.
     if (error == kUnityAdsErrorShowError) {
-      if (_interstitialConnector) {
-        [_interstitialConnector adapterWillDismissInterstitial:self];
-        [_interstitialConnector adapterDidDismissInterstitial:self];
-      } else {
-        [_rewardBasedVideoAdConnector adapterDidCloseRewardBasedVideoAd:self];
+      if (strongNetworkConnector) {
+        [strongNetworkConnector adapterWillDismissInterstitial:self];
+        [strongNetworkConnector adapterDidDismissInterstitial:self];
       }
     }
     return;
   }
 
   NSError *errorWithDescription = GADUnityErrorWithDescription(message);
-  if (_interstitialConnector) {
-    [_interstitialConnector adapter:self didFailAd:errorWithDescription];
-  } else if (_rewardBasedVideoAdConnector) {
-    [_rewardBasedVideoAdConnector adapter:self
-        didFailToLoadRewardBasedVideoAdwithError:errorWithDescription];
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapter:self didFailAd:errorWithDescription];
   }
   _isLoading = NO;
+}
+
+- (void)unityAdsDidStart:(nonnull NSString *)placementId {
+  // nothing to do
+}
+
+#pragma mark - Unity Banner Delegate Methods
+
+- (void)unityAdsBannerDidClick:(nonnull NSString *)placementId {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterDidGetAdClick:self];
+    [strongNetworkConnector adapterWillLeaveApplication:self];
+  }
+}
+
+- (void)unityAdsBannerDidError:(nonnull NSString *)message {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    NSError *error = GADUnityErrorWithDescription(@"Unity Ads Banner internal error");
+    [strongNetworkConnector adapter:self didFailAd:error];
+  }
+}
+
+- (void)unityAdsBannerDidHide:(nonnull NSString *)placementId {
+  NSLog(@"Unity Ads Banner did hide.");
+}
+
+- (void)unityAdsBannerDidLoad:(nonnull NSString *)placementId view:(nonnull UIView *)view {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+
+  if (strongNetworkConnector) {
+    // To support flexible ad sizes, we need to verify if the returned Banner ad fits in the
+    // ad size we requested, and fail the ad request if it doesn't.
+    GADAdSize unityBannerSize =
+        GADAdSizeFromCGSize(CGSizeMake(view.frame.size.width, view.frame.size.height));
+    GADAdSize closestSize =
+        GADClosestValidSizeForAdSizes(unityBannerSize, @[ NSValueFromGADAdSize(_requestedAdSize) ]);
+
+    if (IsGADAdSizeValid(closestSize)) {
+      _bannerDidLoad = YES;
+      [strongNetworkConnector adapter:self didReceiveAdView:view];
+    } else {
+      NSString *errorDescription = [NSString
+          stringWithFormat:@"The banner size loaded (%@) is not valid for the requested size (%@).",
+                           NSStringFromGADAdSize(unityBannerSize),
+                           NSStringFromGADAdSize(_requestedAdSize)];
+      NSLog(@"%@", errorDescription);
+      NSError *error = GADUnityErrorWithDescription(errorDescription);
+      [strongNetworkConnector adapter:self didFailAd:error];
+    }
+  } else {
+    NSLog(@"ERROR: Network connector for UnityAds banner adapter not found.");
+  }
+}
+
+- (void)unityAdsBannerDidShow:(nonnull NSString *)placementId {
+  NSLog(@"Unity Ads Banner is showing.");
+}
+
+- (void)unityAdsBannerDidUnload:(nonnull NSString *)placementId {
+  NSLog(@"Unity Ads Banner has unloaded.");
 }
 
 @end