+++ adapters/InMobi/CHANGELOG.md
@@ -1,5 +1,20 @@
 # InMobi Adapter for Google Mobile Ads SDK for iOS
 
+## Version 7.2.7.0
+- Verified compatibility with InMobi SDK 7.2.7.
+- Updated the adapter to use the new rewarded API.
+- Now requires Google Mobile Ads SDK version 7.42.2 or higher.
+- Added support for flexible banner ad sizes.
+
+## Version 7.2.4.0
+- Verified compatibility with InMobi SDK 7.2.4.
+
+## Version 7.2.1.0
+- Verified compatibility with InMobi SDK 7.2.1.
+
+## Version 7.2.0.0
+- Verified compatibility with InMobi SDK 7.2.0.
+
 ## Version 7.1.2.0
 - Verified compatibility with InMobi SDK 7.1.2.