+++ adapters/Maio/CHANGELOG.md
@@ -1,5 +1,17 @@
 # maio Adapter for Google Mobile Ads SDK for iOS
 
+## Version 1.4.6.0
+- Verified compatibility with maio SDK version 1.4.6.
+- Updated the adapter to use the new rewarded API.
+- Now requires Google Mobile Ads SDK version 7.42.2 or higher.
+
+## Version 1.4.2.0
+- Verified compatibility with maio SDK version 1.4.2.
+
+## Version 1.4.1.0
+- Verified compatibility with maio SDK version 1.4.1.
+- Fixed an issue where maio adapter might fail if multiple MediaIds are provided from AdMob.
+
 ## Version 1.4.0.0
 - Verified compatibility with maio SDK version 1.4.0.