+++ adapters/MyTarget/CHANGELOG.md
@@ -1,5 +1,19 @@
 # myTarget Adapter for Google Mobile Ads SDK for iOS
 
+## Version 5.0.4.0
+- Verified compatibility with myTarget SDK version 5.0.4.
+- Updated the adapter to use the new rewarded API.
+- Now requires Google Mobile Ads SDK version 7.41.0 or higher.
+
+## Version 5.0.1.0
+- Verified compatibility with myTarget SDK version 5.0.1.
+
+## Version 4.8.9.0
+- Verified compatibility with myTarget SDK version 4.8.9.
+
+## Version 4.8.5.0
+- Verified compatibility with myTarget SDK version 4.8.5.
+
 ## Version 4.8.0.0
 - Verified compatibility with myTarget SDK version 4.8.0.
 
@@ -15,6 +29,7 @@
 
 ## Version 4.7.9.0
 - Verified compatibility with myTarget SDK version 4.7.9.
+- Added support for unified native ad.
 
 ## Version 4.7.8.0
 - Verified compatibility with myTarget SDK version 4.7.8.